+++ src/ciceroscm/concentrations_emissions_handler.py
@@ -213,6 +213,21 @@ class ConcentrationsEmissionsHandler:
         self.carbon_cycle = create_carbon_cycle_model(
             model_type, self.pamset, pamset_carbon
         )
+        # Hot-path caches populated by _build_df_gas_cache (called via
+        # reset_with_new_pams below); declared here so they are recognised
+        # as instance attributes and rebuilt on every reset.
+        self._alpha = {}
+        self._sarf_to_erf = {}
+        self._conc_unit = {}
+        self._tau1 = {}
+        self._beta = {}
+        self._nat_em = {}
+        self._nat_em_co2 = 0.0
+        self._inv_tau2_ch4 = 0.0
+        self._inv_tau3_ch4 = 0.0
+        self._yr0 = 0
+        self._emis_arr = {}
+        self._ref_em_species = {}
         # not really needed, but I guess the linter will complain...
         self.reset_with_new_pams(pamset, pamset_carbon, preexisting=False)
 
@@ -372,6 +387,34 @@ class ConcentrationsEmissionsHandler:
                     self.conc[tracer] = {}
                 self.forc[tracer] = np.zeros(years_tot)
         self.forc["Total_forcing"] = np.zeros(years_tot)
+        self._build_df_gas_cache()
+
+    def _build_df_gas_cache(self):
+        """Cache hot-path df_gas column accesses and emis numpy arrays."""
+        # df_gas column caches (static for the run)
+        self._alpha = self.df_gas["ALPHA"].to_dict()
+        self._sarf_to_erf = self.df_gas["SARF_TO_ERF"].to_dict()
+        self._conc_unit = self.df_gas["CONC_UNIT"].to_dict()
+        self._tau1 = self.df_gas["TAU1"].to_dict()
+        self._beta = self.df_gas["BETA"].to_dict()
+        self._nat_em = self.df_gas["NAT_EM"].to_dict()
+        self._nat_em_co2 = float(self.df_gas["NAT_EM"]["CO2"])
+        self._inv_tau2_ch4 = 1.0 / float(self.df_gas["TAU2"]["CH4"])
+        self._inv_tau3_ch4 = 1.0 / float(self.df_gas["TAU3"]["CH4"])
+        # emis as numpy arrays; index 0 corresponds to nystart (self.years[0])
+        self._yr0 = int(self.years[0])
+        self._emis_arr = {col: self.emis[col].to_numpy() for col in self.emis.columns}
+        # Aerosol reference map built once per run (avoids per-call dict construction)
+        self._ref_em_species = {
+            "SO2": ("SO2", self.pamset["qdirso2"]),
+            "SO4_IND": ("SO2", self.pamset["qindso2"]),
+            "OC": ("OC", self.pamset["qoc"]),
+            "BC": ("BC", self.pamset["qbc"]),
+            "BMB_AEROS": ("BMB_AEROS_OC", self.pamset["qbmb"]),
+            "NMVOC": ("NMVOC", self.pamset["qnmvoc"]),
+            "NH3": ("NH3", self.pamset["qnh3"]),
+            "NOx": ("NOx", self.pamset["qnox"]),
+        }
 
     def calculate_strat_quantities(self, yr, conc):
         """
@@ -478,9 +521,9 @@ class ConcentrationsEmissionsHandler:
             + 0.043
         ) * (np.sqrt(c_ch4) - np.sqrt(c0_ch4))
         # Feedback factor: Smith et al 2018
-        q_co2 = self.df_gas["SARF_TO_ERF"]["CO2"] * q_co2  # + FORC_PERT(yr_ix,trc_ix))
-        q_n2o = self.df_gas["SARF_TO_ERF"]["N2O"] * q_n2o  # + FORC_PERT(yr_ix,trc_ix))
-        q_ch4 = self.df_gas["SARF_TO_ERF"]["CH4"] * q_ch4  # + FORC_PERT(yr_ix,trc_ix))
+        q_co2 = self._sarf_to_erf["CO2"] * q_co2  # + FORC_PERT(yr_ix,trc_ix))
+        q_n2o = self._sarf_to_erf["N2O"] * q_n2o  # + FORC_PERT(yr_ix,trc_ix))
+        q_ch4 = self._sarf_to_erf["CH4"] * q_ch4  # + FORC_PERT(yr_ix,trc_ix))
 
         self.forc["CO2"][yr - yr_0] = q_co2
         self.forc["CH4"][yr - yr_0] = q_ch4
@@ -514,6 +557,8 @@ class ConcentrationsEmissionsHandler:
         """
         yr_0 = self.years[0]
         tracer = "TROP_O3"
+        yr_idx = yr - self._yr0
+        ref_yr_idx = self.pamset["ref_yr"] - self._yr0
         # ALOG(1700.0))  !Concentration in 2010 &
         self.conc[tracer][yr] = (
             30.0
@@ -522,10 +567,10 @@ class ConcentrationsEmissionsHandler:
                 np.log(self.conc["CH4"][yr])
                 - np.log(self.conc_in["CH4"][self.pamset["ref_yr"]])
             )
-            + 0.17 * (self.emis["NOx"][yr] - self.emis["NOx"][self.pamset["ref_yr"]])
-            + 0.0014 * (self.emis["CO"][yr] - self.emis["CO"][self.pamset["ref_yr"]])
+            + 0.17 * (self._emis_arr["NOx"][yr_idx] - self._emis_arr["NOx"][ref_yr_idx])
+            + 0.0014 * (self._emis_arr["CO"][yr_idx] - self._emis_arr["CO"][ref_yr_idx])
             + 0.0042
-            * (self.emis["NMVOC"][yr] - self.emis["NMVOC"][self.pamset["ref_yr"]])
+            * (self._emis_arr["NMVOC"][yr_idx] - self._emis_arr["NMVOC"][ref_yr_idx])
         )
         # RBS101115
         # IF (yr_ix.LT.yr_2010) THEN ! Proportional to TROP_O3 build-up
@@ -553,43 +598,17 @@ class ConcentrationsEmissionsHandler:
         float
             Forcing from aerosol in question for year in question
         """
-        ref_emission_species = {
-            "SO2": ["SO2", self.pamset["qdirso2"]],
-            "SO4_IND": ["SO2", self.pamset["qindso2"]],
-            "OC": ["OC", self.pamset["qoc"]],
-            "BC": ["BC", self.pamset["qbc"]],
-            "BMB_AEROS": ["BMB_AEROS_OC", self.pamset["qbmb"]],
-            "NMVOC": ["NMVOC", self.pamset["qnmvoc"]],
-            "NH3": ["NH3", self.pamset["qnh3"]],
-            "NOx": ["NOx", self.pamset["qnox"]],
-        }
-        # Natural emissions
-        # (after IPCC TPII on simple climate models, 1997)
-        # enat = 42.0 Not used, why is this here?
-        # Aerosol forcing used to be scaled to reference year
-        # No just total forcing change
-        # Only with emission to concentration factors differing
-        # These are held in dictionary
-
-        yr_0 = self.years[0]
+        yr_idx = yr - self._yr0
         q = 0
-        # Natural emissions
-        # (after IPCC TPII on simple climate models, 1997)
-        # enat = 42.0 Not used, why is this here?
-        # Emission in reference year
-        # SO2, SO4_IND, BC and OC are treated exactly the same
-        # Only with emission to concentration factors differing
-        # These are held in dictionary
-        if self.df_gas["ALPHA"][tracer] != 0:
-            q = (self.emis[tracer][yr] - self.emis[tracer][yr_0]) * self.df_gas[
-                "ALPHA"
-            ][tracer]
-        elif tracer in ref_emission_species:
-            em_change = (
-                self.emis[ref_emission_species[tracer][0]][yr]
-                - self.emis[ref_emission_species[tracer][0]][yr_0]
+        if self._alpha[tracer] != 0:
+            q = (
+                self._emis_arr[tracer][yr_idx] - self._emis_arr[tracer][0]
+            ) * self._alpha[tracer]
+        elif tracer in self._ref_em_species:
+            em_tracer, coeff = self._ref_em_species[tracer]
+            q = coeff * (
+                self._emis_arr[em_tracer][yr_idx] - self._emis_arr[em_tracer][0]
             )
-            q = ref_emission_species[tracer][1] * em_change
         return q
 
     def conc2forc(
@@ -653,24 +672,21 @@ class ConcentrationsEmissionsHandler:
                 q = self.calc_aerosol_forcing(yr, tracer)
                 f_aero_mag = f_aero_mag + abs(q)
             elif (
-                tracer in self.df_gas.index
-                and self.df_gas["ALPHA"][tracer] != 0  # pylint: disable=compare-to-zero
+                tracer in self._alpha
+                and self._alpha[tracer] != 0  # pylint: disable=compare-to-zero
             ):
                 q = (
                     (self.conc[tracer][yr] - self.conc[tracer][yr_0])
-                    * self.df_gas["ALPHA"][tracer]
-                    * self.df_gas["SARF_TO_ERF"][tracer]
+                    * self._alpha[tracer]
+                    * self._sarf_to_erf[tracer]
                 )  # +forc_pert
 
             elif tracer == "TROP_O3":
-                q = (
-                    self.tropospheric_ozone_forcing(yr)
-                    * self.df_gas["SARF_TO_ERF"][tracer]
-                )
+                q = self.tropospheric_ozone_forcing(yr) * self._sarf_to_erf[tracer]
             elif tracer == "STRAT_H2O":
                 q = (
                     self.pamset["qh2o_ch4"] * self.forc["CH4"][yr - yr_0]
-                ) * self.df_gas["SARF_TO_ERF"][
+                ) * self._sarf_to_erf[
                     tracer
                 ]  # + FORC_PERT(yr_ix,trc_ix)
             elif tracer == "OTHER":
@@ -758,9 +774,7 @@ class ConcentrationsEmissionsHandler:
         if yr < self.pamset["emstart"]:
             self.conc["CO2"][yr] = self.carbon_cycle.co2em2conc(
                 yr,
-                self.emis["CO2_FF"][yr]
-                + self.emis["CO2_AFOLU"][yr]
-                + self.df_gas["NAT_EM"]["CO2"],
+                self.emis["CO2_FF"][yr] + self.emis["CO2_AFOLU"][yr] + self._nat_em_co2,
                 feedback_dict=feedback_dict,
             )
             self.fill_one_row_conc(yr, avoid=["CO2"])
@@ -769,7 +783,7 @@ class ConcentrationsEmissionsHandler:
 
         for tracer, value_dict in self.conc.items():
             # something something postscenario...?
-            if self.df_gas["CONC_UNIT"][tracer] == "-":
+            if self._conc_unit[tracer] == "-":
                 # Forcing calculated from emissions, should now only be TROP_O3
                 continue
             if tracer == "CO2":
@@ -777,7 +791,7 @@ class ConcentrationsEmissionsHandler:
                     yr,
                     self.emis["CO2_FF"][yr]
                     + self.emis["CO2_AFOLU"][yr]
-                    + self.df_gas["NAT_EM"]["CO2"],
+                    + self._nat_em_co2,
                     feedback_dict=feedback_dict,
                 )
                 continue
@@ -789,18 +803,20 @@ class ConcentrationsEmissionsHandler:
             else:
                 conc_local = self.conc_in[tracer][yr]
 
-            q = 1.0 / self.df_gas["TAU1"][tracer]
+            q = 1.0 / self._tau1[tracer]
 
             if tracer == "CH4":
-                self.df_gas.at[tracer, "NAT_EM"] = self.nat_emis_ch4["CH4"][yr]
+                nat_em = self.nat_emis_ch4["CH4"][yr]
                 q = self.methane_lifetime(q, conc_local, yr)
-            if tracer == "N2O":
-                self.df_gas.at[tracer, "NAT_EM"] = self.nat_emis_n2o["N2O"][yr]
+            elif tracer == "N2O":
+                nat_em = self.nat_emis_n2o["N2O"][yr]
+            else:
+                nat_em = self._nat_em[tracer]
 
             # natural emissions, from gasspamfile
-            emis = self.emis[tracer][yr] + self.df_gas["NAT_EM"][tracer]
+            emis = self._emis_arr[tracer][yr - self._yr0] + nat_em
 
-            point_conc = emis / self.df_gas["BETA"][tracer]
+            point_conc = emis / self._beta[tracer]
             # Rewrote this quite a bit from an original loop,
             # but I think it is mathematically equivalent
             value_dict[yr] = point_conc / q + (conc_local - point_conc / q) * np.exp(-q)
@@ -828,11 +844,19 @@ class ConcentrationsEmissionsHandler:
         ch4_wigley_exp = -0.238
         if self.pamset["lifetime_mode"] == "TAR":
             # 1751 is reference conc in 2000
+            yr_idx = yr - self._yr0
+            yr_2000_idx = 2000 - self._yr0
             dln_oh = (
                 -0.32 * (np.log(conc_local) - np.log(1751.0))
-                + 0.0042 * (self.emis["NOx"][yr] - self.emis["NOx"][2000])
-                - 0.000105 * (self.emis["CO"][yr] - self.emis["CO"][2000])
-                - 0.000315 * (self.emis["NMVOC"][yr] - self.emis["NMVOC"][2000])
+                + 0.0042
+                * (self._emis_arr["NOx"][yr_idx] - self._emis_arr["NOx"][yr_2000_idx])
+                - 0.000105
+                * (self._emis_arr["CO"][yr_idx] - self._emis_arr["CO"][yr_2000_idx])
+                - 0.000315
+                * (
+                    self._emis_arr["NMVOC"][yr_idx]
+                    - self._emis_arr["NMVOC"][yr_2000_idx]
+                )
             )
             q = q * (dln_oh + 1)
 
@@ -841,7 +865,7 @@ class ConcentrationsEmissionsHandler:
         elif self.pamset["lifetime_mode"] == "WIGLEY":
             q = q * (((conc_local / 1700.0)) ** (ch4_wigley_exp))
 
-        q = q + 1.0 / self.df_gas["TAU2"]["CH4"] + 1.0 / self.df_gas["TAU3"]["CH4"]
+        q = q + self._inv_tau2_ch4 + self._inv_tau3_ch4
 
         return q